Split LibraryIndexable to its own file
